requirements.txt
fastapi[all]

google-api-python-client
protobuf
pydantic
requests
firebase-admin